ide: 64-bit pdb debugging
git-svn-id: svn://ultimatepp.org/upp/trunk@7581 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/plugin/ndisasm/lib/disasm.c
@@ -1001,7 +1001,7 @@ static const char * const condition_name[16] = {
 };
 
 int32_t disasm(uint8_t *data, char *output, int outbufsize, int segsize,
-            int32_t offset, int autosync, uint32_t prefer)
+               uint64_t offset, int autosync, uint32_t prefer)
 {
     const struct itemplate * const *p, * const *best_p;
     const struct disasm_index *ix;
